+++ __tests__/store/reducers/index.test.ts
@@ -2,7 +2,7 @@ import { createStore } from "redux";
 import { rootReducer } from "../../../src/state/reducers";
 import { appReducer } from "../../../src/state/reducers/appReducer";
 import { networkReducer } from "../../../src/state/reducers/networkReducer";
-import { messageReducer } from "../../../src/state/reducers/postReducer";
+import { messageReducer } from "../../../src/state/reducers/messageReducer";
 import { submitMessageReducer } from "../../../src/state/reducers/sendPostReducer";
 import { IState } from "../../../src/state/state/IState";
