+++ __tests__/store/reducers/messageReducer.test.ts
@@ -8,7 +8,7 @@ import {
 import {
     initialMessage,
     messageReducer,
-} from "../../../src/state/reducers/postReducer";
+} from "../../../src/state/reducers/messageReducer";
 import { IMessagesState } from "../../../src/state/state/IPostsState";
 import { mock } from "../../helpers/mocks";
