+++ src/api/fetchPosts.ts
@@ -1,4 +1,4 @@
-import { MessageWithDate } from "../state/state/IPost";
+import { MessageWithDate } from "../state/state/IMessage";
 
 export async function fetchMessages(uri: string, fetchFn = fetch) {
     const response = await fetchFn(uri);
