+++ src/state/action-creators/submitMessage.ts
@@ -1,6 +1,6 @@
 import { ActionType } from "../actions/ActionType";
 import { ISubmitMessageRequested } from "../actions/IAction";
-import { IMessage } from "../state/IPost";
+import { IMessage } from "../state/IMessage";
 
 export const submitMessage = (payload: IMessage): ISubmitMessageRequested => ({
     type: ActionType.SubmitMessageRequested,
