+++ src/state/reducers/sendPostReducer.ts
@@ -1,7 +1,7 @@
 import { Reducer } from "redux";
 import { ActionType } from "../actions/ActionType";
 import { SubmitMessageAction } from "../actions/SendPostAction";
-import { ISubmitMessageState } from "../state/ISendPostState";
+import { ISubmitMessageState } from "../state/ISubmitMessageState";
 
 /** exponential backoff for submitting a new message to the server */
 export const submitMessageReducer: Reducer<
