+++ src/state/state/IState.ts
@@ -1,7 +1,7 @@
 import { IAppState } from "./IAppState";
 import { INetworkState } from "./INetworkState";
 import { IMessagesState } from "./IPostsState";
-import { ISubmitMessageState } from "./ISendPostState";
+import { ISubmitMessageState } from "./ISubmitMessageState";
 
 export interface IState {
     messages: IMessagesState;
